requirements.txt
mido
fake-bpy-module-3.4